requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.9
flask-cors>=4.0.0
python-dotenv>=1.0.1
databricks-sdk>=0.20.0
databricks-sql-connector>=3.4.0
pandas>=2.1.0
numpy>=1.24.3
python-dateutil>=2.8.2
pytz>=2023.3
six>=1.16.0
typing-extensions>=4.9.0
werkzeug>=3.0.1
pyyaml>=6.0.1
requests>=2.31.0
GitPython>=3.1.44
pyarrow>=15.0.0
databricks-sqlalchemy>=0.1.0
pydantic>=1.8,<2.8
pydantic[email]>=1.8,<2.8
pydantic-settings>=2.2.1
sqlalchemy>=1.4,<2.1
alembic>=1.11.1
psycopg2-binary>=2.9.10
openai>=1.79.0
mlflow>=2.9.0,<2.14.0
